AI-NativeV3-main/apps/academic-service/pyproject.toml
[project]
name = "platform-academic-service"
version = "0.1.0"
description = "CRUDs del dominio académico (universidades, carreras, comisiones)"
requires-python = ">=3.12,<3.13"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "sqlalchemy>=2.0",
    "redis>=5.0",
    "structlog>=24.4",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27",
    "platform-contracts",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    "casbin>=1.36",
    "casbin-sqlalchemy-adapter>=1.4,<2.0",
    "python-jose[cryptography]>=3.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "testcontainers[postgres,redis]>=4.8",
    "hypothesis>=6.108",
    "platform-test-utils",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/academic_service"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--import-mode=importlib"
markers = [
    "integration: tests que requieren testcontainers",
]
